requirements.txt
streamlit
opencv-python
numpy
Pillow
scikit-learn
scipy
matplotlib
